requirements.txt
flask
flask-sqlalchemy
flask-bcrypt 
python-dotenv
flask-session
redis
flask-cors
datetime